refactor: docstring dreamer.py and split shared model pieces into models.py
Converts dreamer.py's top-of-file `#`-comment preamble (the DreamerV3
techniques explainer) into a proper module docstring, and moves the
reusable world-model building blocks - unimix_categorical,
straight_through_sample, categorical_kl, TwoHotHead, Encoder, Decoder,
RSSM, PriorNet, PostNet - into navix/agents/models.py, alongside PPO's
existing shared network components. WorldModel (which wires these into
an RSSM) and the Actor/Critic heads stay in dreamer.py: their docstrings
and call sites are tightly Dreamer-specific (imagination rollouts,
REINFORCE), unlike the pieces that moved.

No public API changes: navix.agents.dreamer re-exports the moved names
via its new `from .models import ...`, and navix/agents/__init__.py's
exports are untouched. tests/test_dreamer.py now imports the moved
names from navix.agents.models directly, matching the new module
structure.
